Update weights on certain headings
There were conflicting weights on the IDL site; the brand team likes to use light weight on black backgrounds and regular weight on light backgrounds for the same size fonts. We have to pick a lane and stick with the regular weight for $expressive-heading-04 (all steps) $expressive-heading-05 (step 7) and $productive-heading-04. 

Basically reverting back to what we had before... apologize for the confusion.